+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.13.0] - 2025-05-19
+## [3.13.0] - 2025-06-02
 * Extend resource group support:
   * Delete `r8s application policies` command group
   * Add `r8s parent group` cli group:
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * `add_resource_group` - Adds specified resource groups to configuration
     * `remove_resource_group` - Removes specified resource groups from configuration
 * Delete `Licenses` collection and related crud commands, move all license-related data to RIGHTSIZER_LICENSES application
+* Add `r8s shape price sync` command to update shape prices for specific AWS region / OS
 
 ## [3.12.0] - 2025-03-14
 * Add ability to run Syndicate RightSizer instance with SaaS ModularSDK / S3